v1.20.24 update
修正NbtShort#getAsNumber方法将意外返回long的问题。

+++ src/main/java/pers/neige/neigeitems/libs/bot/inker/bukkit/nbt/NbtNumeric.java
@@ -56,7 +56,7 @@ public abstract class NbtNumeric<NMS extends RefNbtNumber> extends Nbt<NMS> impl
         } else if (this instanceof NbtLong) {
             return delegate.asLong();
         } else if (this instanceof NbtShort) {
-            return delegate.asLong();
+            return delegate.asShort();
         } else {
             throw new IllegalStateException("Unknown implementation for numeric found: " + this.getClass());
         }
